Update rsw_math and crsw_math, don't pay for doubles when you want floats

+++ glcommon/rsw_math/cmat234.h
@@ -262,17 +262,17 @@ void mult_m4_m4(mat4 c, mat4 a, mat4 b);
 RSW_INLINE void load_rotation_m2(mat2 mat, float angle)
 {
 #ifndef ROW_MAJOR
-	mat[0] = cos(angle);
-	mat[2] = -sin(angle);
+	mat[0] = cosf(angle);
+	mat[2] = -sinf(angle);
 
-	mat[1] = sin(angle);
-	mat[3] = cos(angle);
+	mat[1] = sinf(angle);
+	mat[3] = cosf(angle);
 #else
-	mat[0] = cos(angle);
-	mat[1] = -sin(angle);
+	mat[0] = cosf(angle);
+	mat[1] = -sinf(angle);
 
-	mat[2] = sin(angle);
-	mat[3] = cos(angle);
+	mat[2] = sinf(angle);
+	mat[3] = cosf(angle);
 #endif
 }
